switch the doc of collapse to doc.rs
